requirements.txt
groq
streamlit
rich
python-dotenv
plotly
pdfplumber
